backend/requirements.txt
flask==2.1.0
flask-cors==3.0.10
google-generativeai==0.3.1
python-dotenv==0.20.0
gunicorn==20.1.0
